Add files via upload
Hi Alasdair,

here are my graphs for session two.

In the first graph, I'd like to have the actual percentages on the y-axis. Currently, it just shows the proportion (e.g. 0.8 instead of 80%). How can I change that?
The second graph is, especially in the lower fare scope, quite confusing. Do you have an idea to fix that without splitting it into 3 graphs?

Thanks in advance!
Fritz